Move dataset-difference into utils (public).
