+++ lib/ncdata/utils/__init__.py
@@ -1,5 +1,6 @@
 """General user utility functions."""
 
+from ._compare_nc_datasets import dataset_differences
 from ._save_errors import save_errors
 
-__all__ = ["save_errors"]
+__all__ = ["dataset_differences", "save_errors"]
