+++ tests/integration/test_iris_load_and_save_equivalence.py
@@ -11,7 +11,7 @@ import iris
 import pytest
 
 from ncdata.netcdf4 import from_nc4, to_nc4
-from tests._compare_nc_datasets import dataset_differences
+from ncdata.utils import dataset_differences
 from tests.data_testcase_schemas import session_testdir, standard_testcase
 from tests.integration.equivalence_testing_utils import (
     adjust_chunks,
