+++ tests/unit/netcdf/test_to_nc4.py
@@ -17,7 +17,7 @@ import pytest
 
 from ncdata import NcData
 from ncdata.netcdf4 import from_nc4, to_nc4
-from tests._compare_nc_datasets import dataset_differences
+from ncdata.utils import dataset_differences
 from tests.data_testcase_schemas import make_testcase_dataset
