+++ tests/unit/tests/test_compare_nc_datasets__mainfunctions.py
@@ -11,7 +11,7 @@ import numpy as np
 import pytest
 
 from ncdata import NcAttribute, NcData, NcDimension, NcVariable
-from tests._compare_nc_datasets import dataset_differences
+from ncdata.utils import dataset_differences
 
 # from tests.data_testcase_schemas import _Datatype_Sample_Values, data_types
 # data_types  # avoid 'unused' warning
